refactor: move ParserFactory to cli package
