+++ internal/cli/orchestration.go
@@ -3,11 +3,12 @@ package cli
 import (
 	"fmt"
 	"log"
+	"regexp"
 
 	"github.com/tmknom/actdocs/internal/action"
 	"github.com/tmknom/actdocs/internal/conf"
-	"github.com/tmknom/actdocs/internal/parse"
 	"github.com/tmknom/actdocs/internal/read"
+	"github.com/tmknom/actdocs/internal/util"
 	"github.com/tmknom/actdocs/internal/workflow"
 )
 
@@ -19,7 +20,7 @@ func Orchestrate(source string, formatter *conf.FormatterConfig, sort *conf.Sort
 	}
 	log.Printf("read: %s", source)
 
-	factory := &parse.ParserFactory{Raw: yaml}
+	factory := &ParserFactory{Raw: yaml}
 	parser, err := factory.Factory(formatter, sort)
 	if err != nil {
 		return "", err
@@ -33,14 +34,38 @@ func Orchestrate(source string, formatter *conf.FormatterConfig, sort *conf.Sort
 
 	formatted := ""
 	switch content.(type) {
-	case *parse.ActionAST:
+	case *action.ActionAST:
 		formatter := action.NewActionFormatter(formatter)
-		formatted = formatter.Format(content.(*parse.ActionAST))
-	case *parse.WorkflowAST:
+		formatted = formatter.Format(content.(*action.ActionAST))
+	case *workflow.WorkflowAST:
 		formatter := workflow.NewWorkflowFormatter(formatter)
-		formatted = formatter.Format(content.(*parse.WorkflowAST))
+		formatted = formatter.Format(content.(*workflow.WorkflowAST))
 	default:
 		return "", fmt.Errorf("unsupported AST type: %T", content)
 	}
 	return formatted, nil
+}
+
+type ParserFactory struct {
+	Raw []byte
+}
+
+func (f ParserFactory) Factory(config *conf.FormatterConfig, sort *conf.SortConfig) (util.YamlParser, error) {
+	if f.isReusableWorkflow() {
+		return workflow.NewWorkflowParser(sort), nil
+	} else if f.isCustomActions() {
+		return action.NewActionParser(sort), nil
+	} else {
+		return nil, fmt.Errorf("not found parser: invalid YAML file")
+	}
+}
+
+func (f ParserFactory) isReusableWorkflow() bool {
+	r := regexp.MustCompile(`(?m)^[\s]*workflow_call:`)
+	return r.Match(f.Raw)
+}
+
+func (f ParserFactory) isCustomActions() bool {
+	r := regexp.MustCompile(`(?m)^[\s]*runs:`)
+	return r.Match(f.Raw)
 }
